# src/bkauth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BKAuth API 文档",
        "title": "BKAuth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/oauth/targets": {
            "post": {
                "description": "creates a target with base info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "target"
                ],
                "summary": "target create",
                "operationId": "api-target-create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Target Info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createdTargetSerializer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.targetCreateResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/targets/{target_id}": {
            "put": {
                "description": "updates a target with base info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "target"
                ],
                "summary": "target update",
                "operationId": "api-target-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Target Info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updatedTargetSerializer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.createdTargetSerializer": {
            "type": "object",
            "required": [
                "clients",
                "id",
                "name"
            ],
            "properties": {
                "clients": {
                    "type": "string",
                    "example": "bk_ci,bk_ci1,bk_ci2"
                },
                "description": {
                    "type": "string",
                    "example": "Platform as A Service"
                },
                "id": {
                    "type": "string",
                    "maxLength": 16,
                    "minLength": 3,
                    "example": "bk_ci"
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "example": "BK PaaS"
                }
            }
        },
...
        "handler.targetCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "bk_ci"
                }
            }
        },
...
        "handler.updatedTargetSerializer": {
            "type": "object",
            "properties": {
                "clients": {
                    "type": "string",
                    "example": "bk_ci,bk_ci1,bk_ci2"
                },
                "description": {
                    "type": "string",
                    "example": "Platform as A Service"
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "example": "BK PaaS"
                }
            }
        },
...
        "util.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}